FLIX-198: Customer can not create a requets if previous request in status PAYMENT_WAITING

+++ src/main/java/com/fu/flix/entity/TransactionHistory.java
@@ -23,7 +23,7 @@ public class TransactionHistory {
 
     private String requestCode;
 
-    private Double amount;
+    private Long amount;
 
     private String type;
